@prakashsurya 2018-03-22
@andy-js @gwr @rmustacc Here's another iteration of this change.

What I'm proposing in this version, is introducing a couple new files: `libzfs_taskq.h` and `libzfs_taskq.c`. These files introduce a new taskq implementation that is internal to the `libzfs` library. Thus, with this change, we end up with a taskq implementation in `libfakekernel` which already exists and is used by `libzpool` (and others), and this new taskq implementation in `libzfs` (that's only used by `libzfs`). I didn't make any other changes compared to the prior iterations. Please let me know what y'all think.

I've successfully run this through `zloop` and `zfstest` using our internal Delphix automation.