@pbhenson 2018-03-22
This is great stuff, thanks. We've always left mountpoint set to legacy and used our own custom mounting script to do it in parallel because of the horrid performance for a large number of filesystems.

Any chance you might look at the NFS sharing piece next :)? That also takes forever and a day, so we have sharenfs set to off and again have a custom script to do it.