@prakashsurya 2018-03-22
@pbhenson I agree that it'd be great to also improve the NFS sharing part of the process; unfortunately though, I don't know of any existing plans to do that work.