@andy-js 2018-03-27
I have been doing my own testing with parallel mounts and have observed that enabling SMB shares is a significant bottleneck.  From looking at the code it's not clear to me that this will help.